Spent 3 days basically refactoring the same problem.  Built stand alone Aplollo server to test models, seeds, typeDefs, Schema, Resolvers...and add in the ancillary files.  Issue was not found in our lessons, nor couldl we find inf in online documentation.  We sought help through several avenues and got shut down every time.    Tutor came to rescue.  One line in 4 files each would have had us running 3 days ago.  const { String, Number, Boolean } = mongoose.Schema.Types;  Mongoose need us to identify what types fo types we were using.   const { String, Number, Boolean } = mongoose.Schema.Types;
