+++ server/b4models/Dog.js
@@ -6,6 +6,7 @@ const dogSchema =  new Schema ({
   name: {
     type: String,
     required: true,
+    unique: true,
     trim: true,
   },
   breed: {
