+++ server/b4models/index.js
@@ -1,0 +1,9 @@
+const User = require('./User');
+const Provider = require('./Provider');
+const Dog = require('./Dog');
+const Veterinarian = require('./Veterinarian');
+const Service = require('./Service');
+
+module.exports = { User, Provider, Dog, Veterinarian, Service };
+
+
