+++ server/models/Provider.js
@@ -1,18 +1,15 @@
 const mongoose =require ('mongoose');
 const { Schema } = mongoose;
-const bcrypt = require("bcrypt");
 
 const providerSchema = new Schema({
-   name: {
+  name: {
     type: String,
-    required: true,
-    unique: true,
+    required: true,   
     trim: true,
   },
   email: {
     type: String,
     required: true,
-    unique: true,
     match: [/.+@.+\..+/, 'Must use a valid email address'],
   },
   password: {
@@ -28,14 +25,10 @@ const providerSchema = new Schema({
   address: {
     type: String,
     required: true,
-    unique: true,
+   
   },
-  veterinarian: [
-    {
-      type: Schema.Types.ObjectId,
-      ref: "Veterinarian",
-    },
-  ],
+
+
   about: {
     type: String,
     trim: true,
@@ -48,20 +41,10 @@ const providerSchema = new Schema({
 }
 );
 
-providerSchema.pre('save', async function (next) {
-  if (this.isNew || this.isModified('password')) {
-    const saltRounds = 10;
-    this.password = await bcrypt.hash(this.password, saltRounds);
-  }
 
-  next();
-});
 
-// compare the incoming password with the hashed password
-providerSchema.methods.isCorrectPassword = async function (password) {
-  return bcrypt.compare(password, this.password);
-};
+
  
-const Provider = mongoose.model("Provider", providerSchema);
+const Provider = mongoose.model('Provider', providerSchema);
 
 module.exports = Provider;
